+++ src/ChargeTransport.jl
@@ -23,6 +23,8 @@ using LessUnitful: @local_unitfactors, @ufac_str, @ph_str
 using Printf: @printf
 # for interpolation of data
 using Roots: Roots, find_zero
+# for using in computation of distribution functions
+using SpecialFunctions: erfc, erfcinv
 # PDE solver with a FVM spatial discretization
 using VoronoiFVM: VoronoiFVM, ContinuousQuantity, DiscontinuousQuantity,
     TestFunctionFactory, boundary_dirichlet!, fbernoulli_pm, physics!,
@@ -49,7 +51,7 @@ export tiny_penalty_value
 
 include("ct_distributions.jl")
 
-export Boltzmann, Blakemore, FermiDiracMinusOne, FermiDiracOneHalfBednarczyk
+export Boltzmann, Blakemore, FermiDiracMinusOne, FermiDiracOneHalfBednarczyk, GaussFermiPaasch, GaussFermiSimpson13
 export FermiDiracOneHalfTeSCA
 ##################################################################
 
@@ -98,7 +100,7 @@ export Params, ParamsNodal, ParamsOptical, Data, System
 export BulkRecombination, set_bulk_recombination
 
 export enable_ionic_carrier!
-export enable_trap_carrier!
+export enable_trap_carrier!, constructGaussFermiSimpson13!
 
 export equilibrium_solve!
 export enable_species!, enable_boundary_species!
@@ -123,7 +125,6 @@ export plot_densities, plot_energies, plot_doping, plot_electroNeutralSolutionBo
 export plot_solution, plot_IV
 export plot_densities!, plot_energies!, plot_doping!, plot_electroNeutralSolutionBoltzmann!
 export plot_solution!, plot_IV!
-
 #################################################################
 include("ct_io.jl")
 export read_diodat