+++ src/ct_distributions.jl
@@ -1,5 +1,3 @@
-using ForwardDiff
-
 """
 $(TYPEDSIGNATURES)
 
@@ -90,6 +88,73 @@ function degenerateLimit(x)
     return x < 0 ? NaN : 4 / (3 * sqrt(pi)) * x^(3 / 2)
 end
 
+### Approximation of Gauss-Fermi integral using Paasch's method [J. Appl. Phys. 107, 104501 (2010)]
+### Paasch-Scheinert functions
+"""
+$(TYPEDSIGNATURES)
+
+Paasch-Scheinert approximation of Gauss-Fermi integral parameterized by s = sigma/k_BT (the gaussian width in units of thermal energy).
+For details on function see Paasch and Scheinert, J. Appl. Phys 107, 104501 (2010)
+"""
+struct GaussFermiPaasch{T} <: Function
+    s::T
+end
+function (GaussFermiPaasch::GaussFermiPaasch{T})(x::Real) where {T}
+    function H(s::Real)
+        return sqrt(2) / s * erfcinv(exp(-(s^2) / 2))
+    end
+    function K(s::Real)
+        return 2 * (1 - H(s) / s * sqrt(2 / pi) * exp(1 / 2 * s^2 * (1 - H(s)^2)))
+    end
+    s = GaussFermiPaasch.s
+    if (abs(x) > s^2)
+        G = exp((s * s / 2 - abs(x))) / (1.0 + exp(K(s) * (s * s - abs(x))))
+    else
+        G = 0.5 * erfc(abs(x) / (s * sqrt(2)) * H(s))
+    end
+    if (x > 0)
+        G = 1 - G
+    end
+    return G
+end
+"""
+$(TYPEDEF)
+
+Struct containing information for numerical integration of Gauss-Fermi integral. 
+Can be constructed automatically from ctsys.data using constructGaussFermiSimpson13(data, itrap::Int64).
+"""
+# Approximate Gauss-Fermi integral using Simpsons rule
+struct GaussFermiSimpson13 <: Function
+    s::Float64
+    Et::Float64
+    kBT::Float64
+    nPoints::Int64
+end
+"""
+$(TYPEDSIGNATURES)
+
+Simpson's 1/3 quadrature rule approximation of Gauss-Fermi integral parameterized by s = sigma/k_BT (the gaussian width in units of thermal energy).
+"""
+function (GaussFermiSimpson13::GaussFermiSimpson13)(x::Real)
+
+    @inline function gaussfermi_G(E::Float64, sigma::Float64, Et::Float64, kBT::Float64, x::Real)
+        return exp(-(E - Et) * (E - Et) / (2 * sigma * sigma)) * FermiDiracMinusOne(x - (E - Et) / kBT)
+    end
+
+    s = GaussFermiSimpson13.s
+    Et = GaussFermiSimpson13.Et
+    kBT = GaussFermiSimpson13.kBT
+    nP = Int64(2 * ceil((GaussFermiSimpson13.nPoints) / 2.0) + 1)
+
+    Elower = Et - 6 * s
+    Eupper = Et + 6 * s
+
+    dE = (Eupper - Elower) / (nP - 1.0)
+
+    v = (2.0 * sum(2.0 * gaussfermi_G(E, s, Et, kBT, x) + gaussfermi_G(E + dE, s, Et, kBT, x) for E in (Elower + dE):(2 * dE):(Eupper - 2 * dE)) + gaussfermi_G(Elower, s, Et, kBT, x) + gaussfermi_G(Eupper, s, Et, kBT, x)) * dE / (3.0 * sqrt(2.0π) * s)
+
+    return v
+end
 """
 $(TYPEDSIGNATURES)
 
@@ -100,7 +165,7 @@ function plotDistributions(; Plotter = nothing)
 
     Plotter.close()
 
-    x = -5:0.1:700
+    x = -5:0.1:10
 
     Plotter.semilogy(x, FermiDiracOneHalfBednarczyk.(x), label = "\$F_{1/2}  \$ (Bednarczyk)")
     Plotter.semilogy(x, FermiDiracOneHalfTeSCA.(x), label = "\$F_{1/2} \$ (TeSCA)")
@@ -108,6 +173,10 @@ function plotDistributions(; Plotter = nothing)
     Plotter.semilogy(x, ones(size(x)) / 0.27, "--", label = "\$1/\\gamma=3.\\overline{703}\$", color = (0.6, 0.6, 0.6, 1))
     Plotter.semilogy(x, Blakemore.(x), label = "Blakemore (\$\\gamma=0.27\$)")
     Plotter.semilogy(x, degenerateLimit.(x), label = "degenerate limit")
+    Plotter.semilogy(x, GaussFermiPaasch(1).(x), label = "Gauss-Fermi (Paasch), ŝ = 1")
+    Plotter.semilogy(x, GaussFermiSimpson13(1, 0, 1, 1000).(x), label = "Gauss-Fermi (Simpson 1/3), ŝ = 1", linestyle = "dotted")
+    Plotter.semilogy(x, GaussFermiPaasch(10).(x), label = "Gauss-Fermi (Paasch), ŝ = 10")
+    Plotter.semilogy(x, GaussFermiSimpson13(10, 0, 1, 1000).(x), label = "Gauss-Fermi (Simpson 1/3), ŝ = 10", linestyle = "dotted")
 
     Plotter.xlabel("\$\\eta\$")
     Plotter.ylabel("\$\\mathcal{F}(\\eta)\$")
@@ -141,6 +210,18 @@ function plotDiffusionEnhancements(; Plotter = nothing)
 
     f = ChargeTransport.degenerateLimit; df = x -> ForwardDiff.derivative(f, x)
     Plotter.semilogy(x, f.(x) ./ df.(x), label = "degenerate limit")
+
+    f = ChargeTransport.GaussFermiPaasch(1); df = x -> ForwardDiff.derivative(f, x)
+    Plotter.semilogy(x, f.(x) ./ df.(x), label = "Gauss-Fermi (Paasch), ŝ = 1")
+
+    f = ChargeTransport.GaussFermiSimpson13(1, 0, 1, 1000); df = x -> ForwardDiff.derivative(f, x)
+    Plotter.semilogy(x, f.(x) ./ df.(x), label = "Gauss-Fermi (Simpson 1/3), ŝ = 1", linestyle = "dotted")
+
+    f = ChargeTransport.GaussFermiPaasch(10); df = x -> ForwardDiff.derivative(f, x)
+    Plotter.semilogy(x, f.(x) ./ df.(x), label = "Gauss-Fermi (Paasch), ŝ = 10")
+
+    f = ChargeTransport.GaussFermiSimpson13(10, 0, 1, 1000); df = x -> ForwardDiff.derivative(f, x)
+    Plotter.semilogy(x, f.(x) ./ df.(x), label = "Gauss-Fermi (Simpson 1/3), ŝ = 10", linestyle = "dotted")
 
     Plotter.xlabel("\$\\eta\$")
     Plotter.ylabel("\$g(\\eta)\$")