+++ test/runtests.jl
@@ -22,6 +22,9 @@ end
 
 modname(fname) = splitext(basename(fname))[1]
 
+# Test functions
+include("test_units.jl")
+
 #
 # Include all Julia files in `testdir` whose name starts with `prefix`,
 # Each file `prefixModName.jl` must contain a module named