+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -----------------------------------------------------------------------------------------------
+## v1.8.0
+
+### Added
+- statistics functions to approximate the solution to Gauss-Fermi integrals, `GaussFermiPaasch` and `GaussFermiSimpson13`
+- basic unit tests for statistics functions
+
+### Changed
+- the trap reaction function to a mathematically equivalent formalism which is more flexible and numerically more stable
+- name of abstract type `SingleStateTrap` to `TrapCaptureEscape`
 
 ## v1.7.1